test_pptx_processor and moving copy_file pytest fixture to file_test
